refactor: Update ChatMessage component to use @uiw/react-markdown-preview
The ChatMessage component has been refactored to use the "@uiw/react-markdown-preview" library for rendering markdown content. This change includes updating the import statements and replacing the usage of the "@uiw/react-md-editor" component with the new "@uiw/react-markdown-preview" component. By using this library, the component now provides a better rendering experience for markdown-formatted messages in the chat.